lambdas/get_results_from_winner/requirements.txt
awswrangler==2.20.1
pyarrow==10.0.1
boto3==1.24.59
pandas==1.4.3
requests==2.28.1
python-dotenv==0.21.0 